권한이 있는 사람만 삭제 구현

+++ src/main/java/com/kmg/exam/demo/controller/UsrArticleController.java
@@ -2,6 +2,8 @@ package com.kmg.exam.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -37,8 +39,20 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData<Article> doAdd(String title, String body) {
-		if (Ut.empty(title)) {
+	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if(httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if(isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+		
+		if (Ut.empty(title)) { 
 			return ResultData.from("F-1", "title(을)를 입력해주세요.");
 		}
 
@@ -46,7 +60,7 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "body(을)를 입력해주세요.");
 		}
 
-		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body);
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId, title, body);
 		int id = writeArticleRd.getData1();
 
 		Article article = articleService.getArticle(id);
@@ -55,12 +69,29 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
-			ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
+			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 		}
+
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", "권한이 없습니다.");
+		}
+		
 		articleService.deleteArticle(id);
 
 		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제하였습니다.", id), id);
